refactor(api): move route-table beside the other http helpers
`lib/` in this app means Firebase SDK wiring: singletons and thin
wrappers around an external client. Route-table introspection is neither
— it is two pure functions over a Hono data structure, the same shape as
`http/retry-after.ts`.
