+++ apps/api/src/routes/root.ts
@@ -4,7 +4,7 @@
 import type { Env, Hono as HonoApp } from 'hono';
 import { Hono } from 'hono';
 
-import { isRouteHandler } from '@/lib/hono/route-table.js';
+import { isRouteHandler } from '@/http/route-table.js';
 import type { NegotiatedResponseContentVariables } from '@/middleware/negotiate-content.js';
 import { negotiateContent } from '@/middleware/negotiate-content.js';
 import { rootGetResponseV1 } from '@/profiles/json-schema/root/get-response-v1.js';
